refactor(sentinels): promote seams gate into the sentinels family + repair migration
Last session began generalizing the Seam Observatory into a reusable
`babylon.sentinels` family (shared `base.run_sensor` two-tier runner +
`_ast` static-analysis helpers), moving `babylon.seams` -> `babylon.sentinels.seam`
and `tools/seam_registry_check.py` -> `tools/sentinel_check.py <sensor> --check`.
That migration was left half-applied, so the fast gate was RED:

- `.mise.toml` `check:seams` still shelled the deleted `seam_registry_check.py`;
- `tests/unit/sentinels/` still imported the removed `babylon.seams` module and
  the old tool via a sys.path hack.

This lands the family skeleton and repairs the wiring: `check:seams` now calls
`sentinel_check.py seam --check`; the tests import the check logic from the
package (`babylon.sentinels.seam.checks` + `_ast`/`base`) with no sys.path hack;
stale docstrings/comments updated. Layer-0.5 import boundary enforced by the
renamed import-linter contract (6 kept, 0 broken).

Verified: 16 sentinel tests green, `mise run check:seams` exit 0, ruff clean.
