--- tests/unit/seams/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for the Seam Observatory (``babylon.seams`` + ``tools/seam_*_check.py``)."""
