+++ tests/unit/sentinels/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the Seam Observatory (``babylon.sentinels.seam`` + ``tools/sentinel_check.py``)."""
